CLN a little more clean up in the files
